% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_plot_all.R
\name{hex_plot_all}
\alias{hex_plot_all}
\title{hex plot all}
\usage{
hex_plot_all(df, col_x = NULL, col_fill = NULL)
}
\arguments{
\item{df}{A dataframe from hex_plt()}

\item{col_x}{the column to be plotted on the x-axis}

\item{col_fill}{the column to determine the fill color of points}
}
\value{
a list of hex plots
}
\description{
hex plot all attributs of a dataset or dataframe
}
