% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot_all.R
\name{scatter_plot_all}
\alias{scatter_plot_all}
\title{scatte plot all}
\usage{
scatter_plot_all(df, col_x = NULL, col_color = NULL,
  col_shape = NULL, col_size = NULL)
}
\arguments{
\item{df}{A dataframe from sctr_plt()}

\item{col_x}{the column to be plotted on the x-axis}

\item{col_color}{the column to determine color of points}

\item{col_shape}{the column to be determine shape of points}

\item{col_size}{the column to be determine size of points}
}
\value{
a list of scatter plots
}
\description{
scatter plot all attributs of a dataset or dataframe
}
