% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{sctr_plt}
\alias{sctr_plt}
\title{Scatter Plotting all columns of a dataframe}
\usage{
sctr_plt(df, column_x, column_y, column_color = NULL,
  column_shape = NULL, column_size = NULL, title = paste(column_x,
  "vs", column_y))
}
\description{
Scatter Plotting all columns of a dataframe
}
